Incorporate latest changes to HEAD

+++ crates/fuel-core/src/service/adapters/executor.rs
@@ -56,7 +56,7 @@ impl fuel_core_executor::ports::TransactionsSource for TransactionsSource {
                 minimal_gas_price: self.minimum_gas_price,
                 max_gas: gas_limit,
                 maximum_txs: transactions_limit,
-                maximum_block_size: block_transaction_size_limit,
+                maximum_block_size: block_transaction_size_limit as u64,
                 excluded_contracts: HashSet::default(),
             })
             .unwrap_or_default()
@@ -76,7 +76,7 @@ impl fuel_core_parallel_executor::ports::TransactionsSource for TransactionsSour
         &mut self,
         gas_limit: u64,
         tx_count_limit: u16,
-        block_transaction_size_limit: u32,
+        block_transaction_size_limit: u64,
         filter: Filter,
     ) -> TransactionSourceExecutableTransactions {
         let transactions = self
